+++ server/packages/happy-wire/src/index.ts
@@ -13,3 +13,4 @@ export * from './agentQueue';
 export * from './machineSession';
 export * from './grokPermissionMode';
 export * from './credentialManager';
+export * from './devicePairing';